Fix broken build.
