+++ packages/crumbs-core/managers/themes.ts
@@ -1,7 +1,7 @@
 import CONSTANTS from '../constants';
 import Theme from '../models/theme';
 import type Themes from '../models/themes';
-import { toKebabCase } from '../misc/normalizers';
+import { toCamelCase, toKebabCase } from '../misc/normalizers';
 
 class ThemesManager {
   private themes: Themes;
@@ -68,7 +68,7 @@ class ThemesManager {
   }
 
   getColorVar(key: string): string {
-    if (this.getCurrentTheme()[key]) {
+    if (this.getCurrentTheme()[toCamelCase(key)]) {
       return this.buildColorVar(this.activeThemeName, key);
     }
